Added ability to align with tophat

+++ scripts/create_alignment_script.py
@@ -29,7 +29,7 @@ import re
 import stat
 import sys
 from optparse import OptionParser
-from subprocess import Popen, PIPE
+from subprocess import Popen, PIPE,check_output
 
 import chipsequtil
 from chipsequtil import get_global_settings, get_local_settings, check_org_settings, GLOBAL_SETTINGS_FN, LOCAL_SETTINGS_FN, get_global_tool_settings, get_local_tool_settings, check_tool_settings, GLOBAL_TOOL_SETTINGS_FN, LOCAL_TOOL_SETTINGS_FN
@@ -49,10 +49,9 @@ parser = OptionParser(usage=usage,description=description,epilog=epilog)
 
 script_template = """\
 #!/bin/bash
-source /etc/profile.d/modules.sh
 
 # modules
-module load shhuang bowtie/0.12.7bin bowtie2/2.1.0bin samtools/0.1.18
+%(load_modules)s
 
 # required parameters for the pipeline
 ALIGNER=%(aligner)s
@@ -119,6 +118,7 @@ if __name__ == '__main__' :
         # this dictionary will be used to generate a JSON formatted file with
         # all the relevant settings for the pipeline
         json_dict = {}
+        modules_list = []
         
         ############################################################################
         # get settings
@@ -181,6 +181,8 @@ files in this directory to BAM format and move to a common location on the NFS."
         #####################################
         announce('samtools index configuration')
         all_samtools_settings = all_tool_settings['samtools']
+        modules_list.append(all_samtools_settings['modulefile'])
+
         valid_samtools_ind_list = all_samtools_settings.get('indices','').split()
         samtools_ind_text = """\
 Below are the samtools indices available on this system.  The pipeline will
@@ -232,13 +234,12 @@ correspond to files you need you may add your own to %(local_tool)s.  See %(glob
 
         json_dict['aligner'] = aln
 
-        aln_output_format = 'alignment-sam'
-        aln_output_suffix = '.sam'
-        pipeline_args.append(('--output-format',aln_output_suffix))
-
         ###### bowtie 1 & 2 #####
         if (aln in ['bowtie','bowtie2']):
+            aln_output_suffix = '.sam'
+            pipeline_args.append(('--output-format',aln_output_suffix))
             all_aln_settings = all_tool_settings[aln]
+            modules_list.append(all_aln_settings['modulefile'])
 
             announce('%s options configuration'%aln)
             valid_aln_opt_settings = all_aln_settings.get('opt_sets','').split()
@@ -262,6 +263,14 @@ correspond to files you need you may add your own to %(local_tool)s.  See %(glob
                 print
             json_dict['%s options'%(aln)] = aln_opt
             pipeline_args.append(('--%s-opts'%(aln),all_tool_settings[aln][aln_opt]))
+            aln_output_format = aln_opt
+
+            aln_opt_val_list = all_tool_settings[aln][aln_opt].split(' ')
+            if (('--sam' in aln_opt_val_list) or ('-S' in aln_opt_val_list)):
+                aln_output_suffix = '.sam'
+            else:
+                aln_output_suffix = '.bowtie'
+            pipeline_args.append(('--output-format',aln_output_suffix))
 
             announce('%s index configuration'%aln)
             valid_aln_ind_list = all_aln_settings.get('indices','').split()
@@ -270,7 +279,7 @@ Below are the %(aln)s alignment indices available on this system.  The pipeline 
 use the settings for the entire execution. If you do not see a set of settings that 
 correspond to files you need you may add your own to %(local_tool)s.  See %(glob_tool)s for details.
 """
-            print textwrap.fill(aln_opt_text%{'aln':aln,'local_tool':LOCAL_TOOL_SETTINGS_FN,'glob_tool':GLOBAL_TOOL_SETTINGS_FN},break_long_words=False)
+            print textwrap.fill(aln_ind_text%{'aln':aln,'local_tool':LOCAL_TOOL_SETTINGS_FN,'glob_tool':GLOBAL_TOOL_SETTINGS_FN},break_long_words=False)
             print
 
             wb('Available settings\n')
@@ -294,6 +303,114 @@ correspond to files you need you may add your own to %(local_tool)s.  See %(glob
                 json_dict['nthread'] = nthread
                 pipeline_args.append(('--%s-aln-nthread'%(aln),nthread))
 
+        elif (aln in ['tophat','tophat2']):
+            all_aln_settings = all_tool_settings[aln]
+            modules_list.append(all_aln_settings['modulefile'])
+
+            if (aln=='tophat'):
+                valid_bowtie_versions = ['bowtie']
+                #tophat_version = check_output(["tophat","-v"]).strip().replace(' ','_')
+            elif (aln=='tophat2'):
+                valid_bowtie_versions = ['bowtie','bowtie2']
+                #tophat_version = check_output(["tophat2","-v"]).strip().replace(' ','_')
+
+            announce('%s options configuration'%aln)
+            valid_aln_opt_settings = all_aln_settings.get('opt_sets','').split()
+            aln_opt_text = """
+Below are the %(aln)s alignment options available on this system.  The pipeline will
+use the settings for the entire execution. If you do not see a set of settings that 
+correspond to files you need you may add your own to %(local_tool)s.  See %(glob_tool)s for details.
+"""
+            print textwrap.fill(aln_opt_text%{'aln':aln,'local_tool':LOCAL_TOOL_SETTINGS_FN,'glob_tool':GLOBAL_TOOL_SETTINGS_FN},break_long_words=False)
+            print
+
+            wb('Available settings\n')
+            aln_opt_sets = [(k,all_aln_settings.get(k,'')) for k in all_aln_settings.get('opt_sets','').split()]
+            for aln_opt, settings in aln_opt_sets :
+                wb(aln_opt.ljust(8))
+                print ':', settings
+            aln_opt = ''
+            
+            while aln_opt not in valid_aln_opt_settings :
+                aln_opt = input('Choose '+aln+' options, one of ('+','.join(valid_aln_opt_settings)+')')
+                print
+            json_dict['%s options'%(aln)] = aln_opt
+            pipeline_args.append(('--%s-opts'%(aln),all_tool_settings[aln][aln_opt]))
+            aln_output_format = aln_opt
+
+            aln_opt_val_list = all_tool_settings[aln][aln_opt].split(' ')
+            if ('--no-convert-bam' in aln_opt_val_list):
+                aln_output_suffix = '.sam'
+            else:
+                aln_output_suffix = '.bam'
+            pipeline_args.append(('--output-format',aln_output_suffix))
+
+            announce('bowtie configuration')
+            bowtie_text = """\
+Below are the bowtie settings available on this system.  The pipeline will
+use the settings for the entire execution. If you do not see a set of settings that 
+correspond to files you need you may add your own to %(local_tool)s.  See %(glob_tool)s for details.
+"""
+            print textwrap.fill(bowtie_text%{'local_tool':LOCAL_TOOL_SETTINGS_FN,'glob_tool':GLOBAL_TOOL_SETTINGS_FN},break_long_words=False)
+            print
+
+            wb('bowtie version\n')
+            bowtie_version = ''
+            while bowtie_version not in valid_bowtie_versions:
+                bowtie_version = input('Choose bowtie version, one of ('+','.join(valid_bowtie_versions)+')')
+                print
+            json_dict['bowtie version'] = bowtie_version
+            pipeline_args.append(('--bowtie-version',bowtie_version))
+
+            announce('%s index configuration'%bowtie_version)
+            all_bowtie_settings = all_tool_settings[bowtie_version]
+            modules_list.append(all_bowtie_settings['modulefile'])
+            valid_bowtie_ind_list = all_bowtie_settings.get('indices','').split()
+            wb('Available settings\n')
+            for bowtie_ind in valid_bowtie_ind_list:
+                wb(bowtie_ind.ljust(8))
+                print ':', all_bowtie_settings.get(bowtie_ind+'.description','No description')
+            bowtie_ind = ''
+
+            while bowtie_ind not in valid_bowtie_ind_list:
+                bowtie_ind = input('Choose '+bowtie_version+' index, one of ('+','.join(valid_bowtie_ind_list)+')')
+                print
+
+            json_dict['bowtie index'] = bowtie_ind
+            species = all_bowtie_settings[bowtie_ind+'.species']
+            json_dict['species'] = species
+            pipeline_args.append(('--bowtie-ind-loc',all_bowtie_settings[bowtie_ind+'.loc']))
+
+            announce('%s transcriptome index configuration'%aln)
+            valid_tx_ind_list = all_aln_settings.get('tx_indices','').split()
+            tx_ind_text = """\
+Below are the %(aln)s transcriptome indices available on this system.  The pipeline will
+use the settings for the entire execution. If you do not see a set of settings that 
+correspond to files you need you may add your own to %(local_tool)s.  See %(glob_tool)s for details.
+"""
+            print textwrap.fill(tx_ind_text%{'aln':aln,'local_tool':LOCAL_TOOL_SETTINGS_FN,'glob_tool':GLOBAL_TOOL_SETTINGS_FN},break_long_words=False)
+            print
+
+            wb('Available settings\n')
+            for tx_ind in valid_tx_ind_list:
+                wb(tx_ind.ljust(8))
+                print ':', all_aln_settings.get(tx_ind+'.description','No description')
+            tx_ind = ''
+
+            while tx_ind not in valid_tx_ind_list:
+                tx_ind = input('Choose '+aln+' transcriptome index, one of ('+','.join(valid_tx_ind_list)+')')
+                print
+
+            json_dict['%s transcriptome index'%(aln)] = tx_ind
+            pipeline_args.append(('--%s-gene-mod'%(aln),all_tool_settings[aln][tx_ind+'.gene_mod']))
+            pipeline_args.append(('--%s-tx-ind-loc'%(aln),all_tool_settings[aln][tx_ind+'.loc']))
+
+            nthread = ''
+            while not re.search('^\d+$',nthread) :
+                nthread = input('How many threads would you like to use for mapping EACH experiment?','2')
+                json_dict['nthread'] = nthread
+                pipeline_args.append(('--%s-aln-nthread'%(aln),nthread))
+
         # end if (aln in ['bowtie','bowtie2'])
 
         ############################################################################
@@ -303,12 +420,14 @@ correspond to files you need you may add your own to %(local_tool)s.  See %(glob
         wb('Reading sample sheet and create symlink for read files if needed\n')
         json_dict['alignment output dir'] = dict()
         json_dict['alignment output file'] = dict()
-        json_dict['reads files'] = dict()
+        json_dict['reads1 files'] = dict()
+        json_dict['reads2 files'] = dict()
         json_dict['reads file format'] = dict()
         dest_dir_pattern = all_tool_settings['NameRule']['dest_dir']
         reads_symlink_pattern = all_tool_settings['NameRule']['reads_symlink']
         samp_ifh = open(samp_fn,'r')
         samp_reader = csv.DictReader(samp_ifh,dialect='excel-tab')
+        samp_reader = (dict((k, v.strip()) for k, v in row.items()) for row in samp_reader)
         for samp_line in samp_reader:
             local_id = samp_line['Local-ID']
             samp_line_outfile = samp_line.copy()
@@ -327,8 +446,10 @@ correspond to files you need you may add your own to %(local_tool)s.  See %(glob
             pipeline_args.append(('--output-file',output_file))
             pipeline_args.append(('--final-dir',final_dir))
 
-            json_dict['reads files'][local_id] = []
-            samp_file_parts = samp_line['reads_file_parts'].split(',')
+            samp_file_parts = [('reads1_file_prefix','reads1 files','--reads1-files',
+                                'R1',samp_line['reads1_file_prefix'],samp_line['reads1_file_parts']),
+                               ('reads2_file_prefix','reads2 files','--reads2-files',
+                                'R2',samp_line['reads2_file_prefix'],samp_line['reads2_file_parts'])]
             samp_line_symlink = samp_line.copy()
             symlink_dir = os.path.join(aln_final_path,dest_dir_pattern%(samp_line_symlink))
             if (os.path.exists(symlink_dir)):
@@ -336,23 +457,29 @@ correspond to files you need you may add your own to %(local_tool)s.  See %(glob
                     raise ValueError('Alignment file destination directory name %s exists but is not a directory'%(dest_dir))
             else:
                 os.makedirs(symlink_dir)
-            for part_id in samp_file_parts:
-                samp_line_symlink['part_id'] = part_id
-                symlink_src = os.path.join(samp_line['reads_dir'],'%s%s%s'%(samp_line['reads_file_prefix'],part_id,samp_line['reads_file_suffix']))
-                symlink_target = os.path.join(symlink_dir,reads_symlink_pattern%(samp_line_symlink))
-                if os.path.exists(symlink_target):
-                    if (os.path.realpath(symlink_target) != os.path.abspath(symlink_src)):
-                        ans = raw_input('Symlink %s in current directory points to %s but you asked for %s, overwrite symbolic link? y/[n]'%(symlink_target,os.path.realpath(symlink_target),os.path.abspath(symlink_src)))
-                        if ans=='y':
-                            os.remove(symlink_target)
+            for rf_pref,rf_json,rf_arg,rf_short,pref_str,parts_str in samp_file_parts:
+                json_dict[rf_json][local_id] = []
+                if (pref_str!=''):
+                    parts_list = parts_str.split(',')
+                    for part_id in parts_list:
+                        samp_line_symlink['part_id'] = '%s_%s'%((rf_short,part_id))
+                        symlink_src = os.path.join(samp_line['reads_dir'],'%s%s%s'%(samp_line[rf_pref],part_id,samp_line['reads_file_suffix']))
+                        symlink_target = os.path.join(symlink_dir,reads_symlink_pattern%(samp_line_symlink))
+                        if os.path.exists(symlink_target):
+                            if (os.path.realpath(symlink_target) != os.path.abspath(symlink_src)):
+                                ans = raw_input('Symlink %s in current directory points to %s but you asked for %s, overwrite symbolic link? y/[n]'%(symlink_target,os.path.realpath(symlink_target),os.path.abspath(symlink_src)))
+                                if ans=='y':
+                                    os.remove(symlink_target)
+                                    os.symlink(symlink_src,symlink_target)
+                        else:
+                            print symlink_src,symlink_target
                             os.symlink(symlink_src,symlink_target)
-                else:
-                    print symlink_src,symlink_target
-                    os.symlink(symlink_src,symlink_target)
-                json_dict['reads files'][local_id].append(symlink_target)
-            pipeline_args.append(('--reads-files',','.join(json_dict['reads files'][local_id])))
+                        json_dict[rf_json][local_id].append(symlink_target)
+                pipeline_args.append((rf_arg,','.join(json_dict[rf_json][local_id])))
             pipeline_args.append(('--reads-format',samp_line['reads_file_suffix']))
             json_dict['reads file format'][local_id] = samp_line['reads_file_suffix']
+        # end for samp_line in samp_reader
+        samp_ifh.close()
 
         # put all the command line utility args in json_dict as its own dict
         json_dict['pipeline args'] = pipeline_args
@@ -365,8 +492,16 @@ correspond to files you need you may add your own to %(local_tool)s.  See %(glob
 
         wb('Creating script...\n')
         script_fn = '%s_aln_pipeline.sh'%run_name
+        if (all_tool_settings['tools']['use_modules']):
+            load_modules = """\
+%(run_init)s
+module load %(modules_list)s
+"""%({'run_init':all_tool_settings['modules']['run_init'],'modules_list':' '.join(modules_list)})
+        else:
+            load_modules = ""
         with open(script_fn,'w') as script_f :
-            script_f.write(script_template%{'species':species,'aligner':aln,'def_args':def_args})
+            script_f.write(script_template%{'species':species,'aligner':aln,
+                                            'load_modules':load_modules,'def_args':def_args})
             os.chmod(script_f.name,stat.S_IRWXU|stat.S_IRWXG|stat.S_IROTH)
 
         print end_text%{'script_fn':script_fn}
